+++ examples/resources/nxos_ospf_area/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ospf_area.example "sys/ospf/inst-[OSPF1]/dom-[VRF1]/area-[0.0.0.10]"